Fix menu_role.role_id type to match legacy roles.id for FK

+++ pos-api-new/database/migrations/2026_02_08_114829_create_menu_role_table.php
@@ -14,14 +14,12 @@ return new class extends Migration
         Schema::create('menu_role', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('menu_id');
-            $table->unsignedBigInteger('role_id');
+            $table->unsignedInteger('role_id');
             $table->timestamps();
 
-            // Foreign key constraints
             $table->foreign('menu_id')->references('id')->on('menus')->onDelete('cascade');
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
 
-            // Unique constraint to prevent duplicate menu-role assignments
             $table->unique(['menu_id', 'role_id']);
         });
     }
